refactor(web): logout を features/sessions に統合しリソース名へ揃える
login/logout をセッションリソースとして features/sessions に集約する。
先に分離した理由（login-form 経由で use-login の swr がレイアウトの Storybook
ブラウザテストに波及し use-sync-external-store の interop が崩れる）は、
storybook プロジェクトの optimizeDeps に swr / use-sync-external-store/shim を
明示追加して事前バンドルすることで解消する（noDiscovery: true のため明示が必要）。
