+++ application/apps/web/src/client/components/ui/layout/landing-header/landing-header.stories.tsx
@@ -3,7 +3,7 @@ import { expect, userEvent } from 'storybook/test'
 import { vi } from 'vitest'
 import LandingHeader from './index'
 
-vi.mock('@/client/features/logout/use-logout', () => ({
+vi.mock('@/client/features/sessions/use-logout', () => ({
   default: vi.fn(() => ({
     handleLogout: vi.fn(),
     isLoading: false,
