--- application/apps/web/src/client/features/logout/index.ts
@@ -1,2 +1,0 @@
-export { default as LogoutButton } from './logout-button'
-export { default as SidebarLogoutButton } from './sidebar-logout-button'
