+++ application/apps/web/src/client/features/sessions/sidebar-logout-button.tsx
