+++ application/apps/web/src/client/features/sessions/use-logout.ts
